Enable the VST2 SDK if we have it, and include the Element plugins.
This enables VST2 support in the host, and also enables building the
Element (VST2) plugins so that Element can be used in other hosts.

TODO: Figure out which version of the VST3 SDK is needed to make the
build of the VST3 Element plugins work on Linux. I've tried both VST3
3.6.10 and the latest 3.7 version, but Element's JUCE version seems to
like neither. So, for the time being, only the VST2 version of the
plugins are included in the package.